+++ src/condor_ckpt/condor_file_compress.h
@@ -44,8 +44,8 @@ public:
 
 	virtual int open(const char *url, int flags, int mode);
 	virtual int close();
-	virtual int read(int offset, char *data, int length);
-	virtual int write(int offset, char *data, int length);
+	virtual int read(off_t offset, char *data, int length);
+	virtual int write(off_t offset, char *data, int length);
 
 	virtual int fcntl( int cmd, int arg );
 	virtual int ioctl( int cmd, long arg );
@@ -58,7 +58,7 @@ public:
 	virtual int	is_writeable();
 	virtual int	is_seekable();
 
-	virtual int	get_size();
+	virtual off_t get_size();
 	virtual char const	*get_url();
 
 	virtual int get_unmapped_fd();
@@ -95,10 +95,10 @@ private:
 	CondorFile *original;
 
 	// Current offset into the physical file
-	int poffset;
+	off_t poffset;
 
 	// Current offset into the virtual (compressed) file
-	int voffset;
+	off_t voffset;
 
 	// Accumulating checksum for writes
 	int crc;
